requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
attrs==25.4.0
blinker==1.9.0
cffi==2.0.0
click==8.1.8
coverage==7.10.7
exceptiongroup==1.3.0
fastapi==0.121.0
Flask==3.1.2
h11==0.16.0
idna==3.11
importlib_metadata==8.7.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mypy==1.18.2
mypy_extensions==1.1.0
numpy==2.3.5
packaging==25.0
pathspec==0.12.1
pluggy==1.6.0
py==1.11.0
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
PyQt5==5.15.11
PyQt5-Qt5==5.15.18
PyQt5_sip==12.17.1
pytest==8.4.2
pytest-cov==7.0.0
pytest-pythonpath==0.7.3
python-multipart==0.0.20
shapely==2.1.2
sniffio==1.3.1
starlette==0.49.3
svgwrite==1.4.3
toml==0.10.2
tomli==2.3.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
Werkzeug==3.1.3
zipp==3.23.0
